+++ R/get_object.R
@@ -42,7 +42,12 @@
 #' @references \href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectHEAD.html}{API Document: HEAD Object}
 #' @seealso \code{\link{get_bucket}}, \code{\link{put_object}}, \code{\link{delete_object}}
 #' @export
-get_object <- function(object, bucket, headers = list(), parse_response = FALSE, ...) {
+get_object <- 
+function(object, 
+         bucket, 
+         headers = list(), 
+         parse_response = FALSE, 
+         ...) {
     if (missing(bucket)) {
         bucket <- get_bucketname(object)
     } 
@@ -59,7 +64,12 @@ get_object <- function(object, bucket, headers = list(), parse_response = FALSE,
 
 #' @rdname getobject
 #' @export
-save_object <- function(object, bucket, file, headers = list(), ...) {
+save_object <- 
+function(object, 
+         bucket, 
+         file, 
+         headers = list(), 
+         ...) {
     if (missing(file)) {
         stop('argument "file" is missing, with no default')
     }